api/requirements.txt
duckdb
requests
python-dotenv
fastapi==0.115.2
starlette==0.40.0
uvicorn
polars
authlib
httpx
titlecase==2.4.1
import boto3